+++ backend/src/main/java/com/example/backend/repository/AnaTabloRepository.java
@@ -3,5 +3,5 @@ package com.example.backend.repository;
 import com.example.backend.model.AnaTablo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface AnaTabloRepository extends JpaRepository<AnaTablo, Long> {
+public interface AnaTabloRepository extends JpaRepository<AnaTablo, Float> {
 }
