+++ backend/src/main/java/com/example/backend/repository/DetTabloRepository.java
@@ -3,5 +3,5 @@ package com.example.backend.repository;
 import com.example.backend.model.DetTablo;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface DetTabloRepository extends JpaRepository<DetTablo, Long> {
+public interface DetTabloRepository extends JpaRepository<DetTablo, Float> {
 }
